refactor(ui): migrate shared page to design system
- Reduces SharedPage from 780 LOC to 218 LOC
- Extracts 4 sub-components into src/components/shared-browser/:
    shared-tab-nav       - Commands/Skills/Agents tab switcher + metric cards
    shared-item-list     - Left browser pane with search, filter, item rows
    shared-item-viewer   - Right content pane with markdown display
    shared-markdown-viewer - Self-contained markdown parser + renderer
- Adopts PageShell + PageHeader + ConfigLayout browser layout
- Preserves all flows, i18n keys, error handling, and data fetching verbatim
