+++ ui/src/components/shared-browser/index.ts
@@ -1,0 +1,4 @@
+export { SharedTabNav, type SharedTabType, type SharedTabMeta } from './shared-tab-nav';
+export { SharedItemList } from './shared-item-list';
+export { SharedItemViewer } from './shared-item-viewer';
+export { SharedMarkdownViewer } from './shared-markdown-viewer';
